aerten-web/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docopt==0.6.2
drf-nested-routers==0.94.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.1
locust==2.33.2
MarkupSafe==3.0.2
model-bakery==1.20.4
msgpack==1.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pyzmq==26.3.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
